+++ packages/clerk-js/src/utils/url.ts
@@ -402,6 +402,7 @@ const frontendApiRedirectPathsWithUserInput: string[] = [
 const frontendApiRedirectPathsNoUserInput: string[] = [
   '/v1/verify', // magic links
   '/v1/tickets/accept', // ticket flow
+  '/oauth/authorize-with-immediate-redirect', // OAuth 2 identity provider
 ];
 
 export function isRedirectForFAPIInitiatedFlow(frontendApi: string, redirectUrl: string): boolean {