+++ packages/tanstack-react-start/src/server/getAuth.ts
@@ -1,29 +1,26 @@
 import type { AuthenticateRequestOptions, GetAuthFn } from '@clerk/backend/internal';
 import { getAuthObjectForAcceptedToken } from '@clerk/backend/internal';
+import type { PendingSessionOptions } from '@clerk/types';
+import { getContext } from '@tanstack/react-start/server';
 
 import { errorThrower } from '../utils';
-import { noFetchFnCtxPassedInGetAuth } from '../utils/errors';
-import { authenticateRequest } from './authenticateRequest';
-import { loadOptions } from './loadOptions';
-import type { LoaderOptions } from './types';
+import { clerkHandlerNotConfigured, noFetchFnCtxPassedInGetAuth } from '../utils/errors';
 
-type GetAuthOptions = { acceptsToken?: AuthenticateRequestOptions['acceptsToken'] } & Pick<LoaderOptions, 'secretKey'>;
+type GetAuthOptions = PendingSessionOptions & Pick<AuthenticateRequestOptions, 'acceptsToken'>;
 
 export const getAuth: GetAuthFn<Request, true> = (async (request: Request, opts?: GetAuthOptions) => {
   if (!request) {
     return errorThrower.throw(noFetchFnCtxPassedInGetAuth);
   }
 
-  const { acceptsToken, ...restOptions } = opts || {};
+  const authObjectFn = getContext('auth');
 
-  const loadedOptions = loadOptions(request, restOptions);
+  if (!authObjectFn) {
+    return errorThrower.throw(clerkHandlerNotConfigured);
+  }
 
-  const requestState = await authenticateRequest(request, {
-    ...loadedOptions,
-    acceptsToken: 'any',
-  });
+  // We're keeping it a promise for now to minimize breaking changes
+  const authObject = await Promise.resolve(authObjectFn({ treatPendingAsSignedOut: opts?.treatPendingAsSignedOut }));
 
-  const authObject = requestState.toAuth();
-
-  return getAuthObjectForAcceptedToken({ authObject, acceptsToken });
+  return getAuthObjectForAcceptedToken({ authObject, acceptsToken: opts?.acceptsToken });
 }) as GetAuthFn<Request, true>;