+++ packages/nextjs/CHANGELOG.md
@@ -1,5 +1,19 @@
 # Change Log
 
+## 6.31.4
+
+### Patch Changes
+
+- Add types to safe-node-apis modules. ([#6597](https://github.com/clerk/javascript/pull/6597)) by [@panteliselef](https://github.com/panteliselef)
+
+- Fix keyless drift detection failure caused by invalid fs module methods. ([#6594](https://github.com/clerk/javascript/pull/6594)) by [@heatlikeheatwave](https://github.com/heatlikeheatwave)
+
+- Updated dependencies [[`e52bf8e`](https://github.com/clerk/javascript/commit/e52bf8ebef74a9e123c69b69acde1340c01d32d7), [`c043c19`](https://github.com/clerk/javascript/commit/c043c1919854aaa5b9cf7f6df5bb517f5617f7a1), [`7bb644a`](https://github.com/clerk/javascript/commit/7bb644ad8a7bf28c6010aad6ae0c36f587529fcc), [`c28d29c`](https://github.com/clerk/javascript/commit/c28d29c79bb4f144d782313ca72df7db91a77340), [`172e054`](https://github.com/clerk/javascript/commit/172e054a3511be12d16ba19037db320c2d9838bf)]:
+  - @clerk/types@4.81.0
+  - @clerk/backend@2.9.4
+  - @clerk/clerk-react@5.43.1
+  - @clerk/shared@3.22.0
+
 ## 6.31.3
 
 ### Patch Changes